Renamed to "Fixture" factory.
